+++ BookStore/src/main/java/lt/judalabiau/BookStore/controllers/UserController.java
@@ -1,0 +1,5 @@
+package lt.judalabiau.BookStore.controllers;
+
+public class UserController {
+
+}
