--- BookStore/src/main/java/lt/judalabiau/BookStore/models/repositories/controllers/BookController.java
@@ -1,5 +1,0 @@
-package lt.judalabiau.BookStore.models.repositories.controllers;
-
-public class BookController {
-
-}
